// business-service/src/enum/keycloak.role.enum.ts
export enum KeycloakRoleEnum {
  FULL_ACCESS_ROLE = 'FullAccessRole',
  PUT_GENRE_ROLE = 'PutGenreRole',
  PUT_CONCERT_ROLE = 'PutConcertRole',
  FULL_ACCESS_ADMIN_ACCOUNT_ROLE = 'FullAccessAdminAccountRole',
  VIEW_GENRE_ROLE = 'ViewGenreRole',
  DELETE_CONCERT_ROLE = 'DeleteConcertRole',
  PUT_ADMIN_ACCOUNT_ROLE = 'PutAdminAccountRole',
  VIEW_CONCERT_ROLE = 'ViewConcertRole',
  FULL_ACCESS_GENRE_ROLE = 'FullAccessGenreRole',
  DELETE_ADMIN_ACCOUNT_ROLE = 'DeleteAdminAccountRole',
  FULL_ACCESS_CONCERT_ROLE = 'FullAccessConcertRole',
  DELETE_GENRE_ROLE = 'DeleteGenreRole',
  CUSTOMER_ROLE = 'CustomerRole',
  VIEW_ADMIN_ACCOUNT_ROLE = 'ViewAdminAccountRole',
}